presented TeleportPosition component and extracted terrain
