+++ Explorer/Assets/DCL/RealmNavigation/TeleportController.cs
@@ -44,41 +44,39 @@ namespace DCL.RealmNavigation
         }
 
         /// <summary>
-        /// If current scene is still loading it will block the teleport until its assets are resolved or timed out
+        ///     If current scene is still loading it will block the teleport until its assets are resolved or timed out
         /// </summary>
         public UniTask<WaitForSceneReadiness?> TeleportToSceneSpawnPointAsync(Vector2Int parcel, AsyncLoadProcessReport loadReport, CancellationToken ct) =>
-            TeleportAsync(parcel, loadReport, false, ct);
-            // TeleportAsync(parcel, TeleportationUtils.PickTargetWithOffset, loadReport, ct);
+            TeleportAsync(parcel, loadReport, ct);
 
         /// <summary>
-        /// Debug teleportation
+        ///     Debug Widget teleportation
         /// </summary>
         public UniTask TeleportToParcelAsync(Vector2Int parcel, AsyncLoadProcessReport loadReport, CancellationToken ct) =>
-            TeleportAsync(parcel, loadReport, true, ct);
-            // TeleportAsync(parcel, TeleportationUtils.PickTarget, loadReport, ct);
+            TeleportAsync(parcel, loadReport, ct, nullifySceneDef: true);
 
-            private async UniTask<WaitForSceneReadiness?> TeleportAsync(Vector2Int parcel, AsyncLoadProcessReport loadReport, bool isFromDebugWindow, CancellationToken ct)
+        private async UniTask<WaitForSceneReadiness?> TeleportAsync(Vector2Int parcel, AsyncLoadProcessReport loadReport, CancellationToken ct, bool nullifySceneDef = false)
+        {
+            if (retrieveScene == null)
             {
-                if (retrieveScene == null)
-                {
-                    world?.AddOrGet(playerEntity, new PlayerTeleportIntent(null, parcel, null, ct, loadReport));
-                    loadReport.SetProgress(1f);
-                    return null;
-                }
+                world?.AddOrGet(playerEntity, new PlayerTeleportIntent(null, parcel, ct, loadReport));
+                loadReport.SetProgress(1f);
+                return null;
+            }
 
-                SceneEntityDefinition? sceneDef = await retrieveScene.ByParcelAsync(parcel, ct);
+            SceneEntityDefinition? sceneDef = await retrieveScene.ByParcelAsync(parcel, ct);
 
-                if (sceneDef != null)
-                    if(isFromDebugWindow || !TeleportationUtils.IsTramLine(sceneDef.metadata.OriginalJson.AsSpan()))
-                        parcel = sceneDef.metadata.scene.DecodedBase; // Override parcel as it's a new target
+            if (sceneDef != null && !TeleportationUtils.IsTramLine(sceneDef.metadata.OriginalJson.AsSpan()))
+            {
+                parcel = sceneDef.metadata.scene.DecodedBase; // Override parcel as it's a new target
 
-                if (isFromDebugWindow)
+                if (nullifySceneDef)
                     sceneDef = null;
-
+            }
 
             await UniTask.Yield(PlayerLoopTiming.PostLateUpdate);
 
-            world?.AddOrGet(playerEntity, new PlayerTeleportIntent(sceneDef, parcel, position: null, ct, loadReport));
+            world?.AddOrGet(playerEntity, new PlayerTeleportIntent(sceneDef, parcel, ct, loadReport));
 
             if (sceneDef == null)
             {
